Split app.go into smaller, focused files for maintainability
The main app file had grown to ~1700 lines, making it difficult to navigate and maintain. This reorganizes the code into logical units: - app.go: Core types, constructor, state management - app_update.go: Main Update function and message routing - app_view.go: View rendering functions - app_helpers.go: Session selection, messaging, git operations - app_mouse.go: Mouse event handling and coordinate adjustment
